# pkg/api/registrar/openapiv2/registrar.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "registrar.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Registrar"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/deactivatenode": {
      "post": {
        "operationId": "Registrar_DeactivateNode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registrarDecativateNodeRequest"
            }
          }
        ],
        "tags": [
          "Registrar"
        ]
      }
    },
    "/v1/getaccount": {
      "post": {
        "operationId": "Registrar_GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/registrarGetAccountResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registrarGetAccountRequest"
            }
          }
        ],
        "tags": [
          "Registrar"
        ]
      }
    },
    "/v1/getnoderegistrations": {
      "get": {
        "operationId": "Registrar_GetNodeRegistrations",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/registrarGetNodeRegistrationsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Registrar"
        ]
      }
    },
    "/v1/registernode": {
      "post": {
        "operationId": "Registrar_RegisterNode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/registrarRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registrarRegisterNodeRequest"
            }
          }
        ],
        "tags": [
          "Registrar"
        ]
      }
    },
    "/v1/registerworkload": {
      "post": {
        "operationId": "Registrar_RegisterWorkload",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/registrarRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registrarRegisterWorkloadRequest"
            }
          }
        ],
        "tags": [
          "Registrar"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "registrarDecativateNodeRequest": {
      "type": "object",
      "properties": {
        "NodeId": {
          "type": "string"
        }
      }
    },
    "registrarGetAccountRequest": {
      "type": "object",
      "properties": {
        "Selectors": {
          "$ref": "#/definitions/registrarSelectorList"
        },
        "NodeType": {
          "$ref": "#/definitions/registrarNodeType"
        }
      }
    },
    "registrarGetAccountResponse": {
      "type": "object",
      "properties": {
        "Account": {
          "type": "string"
        },
        "NodeGroup": {
          "type": "string"
        }
      }
    },
    "registrarGetNodeRegistrationsResponse": {
      "type": "object",
      "properties": {
        "NodeRegistrations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/registrarNodeRegistration"
          }
        }
      }
    },
    "registrarNodeRegistration": {
      "type": "object",
      "properties": {
        "NodeId": {
          "type": "string"
        },
        "NodeType": {
          "$ref": "#/definitions/registrarNodeType"
        },
        "Selectors": {
          "$ref": "#/definitions/registrarSelectorList"
        },
        "NodeGroup": {
          "type": "string"
        },
        "RegisteredBy": {
          "type": "string"
        },
        "AccountUUID": {
          "type": "string"
        },
        "RegisteredAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "registrarNodeType": {
      "type": "string",
      "enum": [
        "TPM",
        "GCP",
        "AWS",
        "AZURE",
        "GITHUB"
      ],
      "default": "TPM"
    },
    "registrarRegisterNodeRequest": {
      "type": "object",
      "properties": {
        "NodeType": {
          "$ref": "#/definitions/registrarNodeType"
        },
        "Selectors": {
          "$ref": "#/definitions/registrarSelectorList"
        },
        "NodeGroup": {
          "type": "string"
        }
      }
    },
    "registrarRegisterResponse": {
      "type": "object",
      "properties": {
        "spiffeID": {
          "type": "string"
        }
      }
    },
    "registrarRegisterWorkloadRequest": {
      "type": "object",
      "properties": {
        "WorkloadType": {
          "$ref": "#/definitions/registrarWorkloadType"
        },
        "Selectors": {
          "$ref": "#/definitions/registrarSelectorList"
        },
        "NodeGroup": {
          "type": "string"
        },
        "WorkloadName": {
          "type": "string"
        }
      }
    },
    "registrarSelector": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "registrarSelectorList": {
      "type": "object",
      "properties": {
        "Selectors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/registrarSelector"
          }
        }
      }
    },
    "registrarWorkloadType": {
      "type": "string",
      "enum": [
        "OCI",
        "BIN"
      ],
      "default": "OCI"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}